Remove meta-device-cloud-whitelist.conf
Issue: LIN1018-1377

The PNWHITELIST is wrlinux specific, so it should be set in in wrlinux layer.

And there is a side effect if set it in meta-device-cloud layer: the
PNWHITELIST_meta-python and PNWHITELIST_networking-layer which is set in this
layer would be seen by all configurations when meta-device-cloud layer is
present, this is not we want.

(LOCAL REV: NOT UPSTREAM) -- Will send to github after local review

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>